Task-07 Description: Refactored
