+++ PokeApp/Podfile
@@ -18,5 +18,8 @@ target 'PokeApp' do
           end
         end
       end
+      installer.pods_project.build_configurations.each do |config|
+        config.build_settings["EXCLUDED_ARCHS[sdk=iphonesimulator*]"] = "arm64"
+      end
     end
 end
