+++ PokeApp/PokeApp/Layers/Builder/AppFlow/PokeListScreenBuilder.swift
@@ -11,6 +11,7 @@ protocol PokeListScreenBuilder {
     static func build(coordinatorDelegate: CommonControllerToCoordinatorDelegate?) -> UIViewController
 }
 
+/// Pokemon listesinin gösterileceği ekranı yaratır ve bağımlılıklarını inject eder.
 struct PokeListScreenBuilderImpl: PokeListScreenBuilder {
     static func build(coordinatorDelegate: CommonControllerToCoordinatorDelegate?) -> UIViewController {
         let vc = PokeListVC(nibName: PokeListVC.className, bundle: nil)
