+++ PokeApp/PokeApp/Layers/Data/Repository/PokeRepository.swift
@@ -21,6 +21,11 @@ struct PokeRepositoryImpl: PokeRepository {
     typealias PokeAbilityCompletionType = (AbilityDetail?, Error?) -> Void
     
     
+    /// Servisten limit ve offset parametreleri ile pokemon listesini döner.
+    /// - Parameters:
+    ///   - limit: Tek istekte kaç adet item getirileceğini belirler. - Int
+    ///   - offset: Listenin hangi indexten başlayacağını belirler. - Int
+    ///   - completion: Pokemon dizisi ve error dönen closure.
     func list(limit: Int, offset: Int, completion: @escaping PokeListCompletionType) {
         APIHandler.shared.processRequest(target: Endpoint.list(limit: limit, offset: offset)).done { (response: PokemonList) in
             completion(response, nil)
@@ -29,6 +34,10 @@ struct PokeRepositoryImpl: PokeRepository {
         }
     }
     
+    /// Servisten belirli id ile pokemon detay verisini döner.
+    /// - Parameters:
+    ///   - id: Pokemon id
+    ///   - completion: Pokemon verisi ve error dönen closure.
     func detail(id: Int, completion: @escaping PokeDetailCompletionType) {
         APIHandler.shared.processRequest(target: Endpoint.detail(id: id)).done { (response: PokemonDetail) in
             completion(response, nil)
@@ -37,6 +46,10 @@ struct PokeRepositoryImpl: PokeRepository {
         }
     }
     
+    /// Servisten belirli pokemana ait yeteneğin detaylarını döner.
+    /// - Parameters:
+    ///   - id: Ability id
+    ///   - completion: AbilityDetail ve Error dönen closure.
     func ability(id: Int, completion: @escaping PokeAbilityCompletionType) {
         APIHandler.shared.processRequest(target: Endpoint.ability(id: id)).done { (response: AbilityDetail) in
             completion(response, nil)
