+++ PokeApp/PokeApp/Utilities/Base/BaseViewModel.swift
@@ -11,6 +11,7 @@ protocol BaseViewModel {
     func start()
 }
 
+// ViewModel ile View arasındaki iletişimi sağlayan clousure içerisine event göndermek için kullanılır.
 enum ObservationType<T, E> {
     case updateUI(data: T? = nil), error(error: E?)
 }
